requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
archspec==0.2.5
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
babel==2.18.0
bcrypt==5.0.0
Beaker==1.12.1
beautifulsoup4==4.14.3
binwalk==2.3.4
bleach==6.3.0
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
blosc2==3.8.0
boltons==25.0.0
boto3==1.42.56
botocore==1.42.56
Bottleneck==1.6.0
Brlapi==0.8.7
Brotli==1.2.0
bs4==0.0.2
cachetools==5.5.2
capstone==6.0.0a5
cbor2==5.8.0
certifi==2025.7.9
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-357-build/cockpit-357/tmp/wheel/cockpit-357-py3-none-any.whl#sha256=474f6d2f07f5e1fc6d7a1714574c3620ecb9769fd645f220ab4b2314f02dc195
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.3
commonmark==0.9.1
conda==25.11.0
conda-libmamba-solver==25.4.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
constantly==23.10.4
contourpy==1.3.3
crispy-bootstrap5==2025.6
crypt_r==3.13.1
cryptography==43.0.3
cson==0.8
cssselect==1.3.0
cupshelpers==1.0
customtkinter==5.2.2
cycler==0.11.0
daphne==4.2.1
darkdetect==0.8.0
dasbus==1.7
datasets==4.7.0
dbus-python==1.4.0
dbus_next==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
Django==5.2.11
django-crispy-forms==2.5
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnf==4.24.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.128.4
fastjsonschema==2.21.2
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.25.2
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2026.1.0
gcsfs==2025.9.0
ggshield==1.47.0
gh==0.0.4
gitdb==4.0.12
GitPython==3.1.46
gmpy2==2.2.2
google-api-core==2.27.0
google-auth==2.45.0
google-auth-oauthlib==1.2.3
google-cloud-core==2.3.3
google-cloud-storage==2.14.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.63.0
gpg==1.24.3
greenlet==3.1.1
grpcio==1.48.4
grpcio-status==1.48.4
gssapi==1.7.3
h11==0.16.0
hf-xet==1.4.2
holehe==1.61
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
humanize==4.13.0
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
Incremental==24.11.0
inkex==1.4.0
intervaltree==3.2.0
invoke==2.2.1
ipaclient==4.13.1
ipalib==4.13.1
ipaplatform==4.13.1
ipapython==4.13.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
iso639==0.1.4
isoduration==20.11.0
itsdangerous==2.2.0
jdcal==1.4.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
jwcrypto==1.4.2
kagglehub==1.0.0
kagglesdk==0.1.16
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
libmambapy==2.3.2
louis==3.33.0
lxml==6.0.1
Mako==1.2.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
menuinst==2.3.1
mercurial==7.1.1
mistune==3.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
multiprocess==0.70.18
mutagen==1.47.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
ndindex==1.10.0
neo4j==6.1.0
nest-asyncio==1.6.0
netaddr==1.3.0
nftables==0.1
notebook==7.5.3
notebook_shim==0.2.4
numexpr==2.12.1
numpy==2.3.5
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.6
Paste==3.10.1
perf==0.1
pexpect==4.9.0
phonenumbers==9.0.21
pid==3.0.4
pillow==11.3.0
platformdirs==3.0.0
pluggy==1.6.0
plumbum==1.10.0
ply==3.11
pooch==1.9.0
productmd==1.50
progressbar2==4.5.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.22.3
protobuf==3.19.6
psutil==7.1.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwntools==4.15.0
pwquality==1.4.5
py-cpuinfo==9.0.0
py-ubjson==0.16.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycairo==1.28.0
pycares==4.10.0
pycosat==0.6.6
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.23.0
pycryptodomex==3.23.0
pycups==2.0.4
pycurl==7.45.6
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
pyelftools==0.32
pyenchant==3.2.2
pygitguardian==1.28.0
Pygments==2.19.1
PyGObject==3.54.5
pyinotify==0.9.6
PyJWT==2.6.0
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
PyMySQL==1.1.2
PyNaCl==1.6.1
pyOpenSSL==25.1.0
pyparsing==3.1.2
pyparted==3.13.0
pypdf==6.7.0
pypng==0.0.21
PyQt5==5.15.12.dev2507081429
PyQt5_sip==12.17.1
pyserial==3.5
PySide6==6.10.2
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==0.21.1
python-json-logger==4.0.0
python-ldap==3.4.5
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.1
python-yubico==1.3.3
pytz==2026.1
pyudev==0.24.3
pyusb==1.3.1
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.16-build/kernel-6.18.16/linux-6.18.16-200.fc43.x86_64/tools/net/ynl
pyzmq==27.1.0
qrcode==8.0
QtPy==2.4.3
RapidFuzz==3.12.2
redis==7.1.1
referencing==0.36.2
regex==2026.2.28
rendercv==2.5
rendercv-fonts==0.5.1
reportlab==4.4.9
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
requests-futures==1.0.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.5.1
ROPGadget==7.7
rpds-py==0.27.0
rpm==6.0.1
rpmautospec==0.8.3
rpmautospec-core==0.1.5
rpyc==6.0.2
rsa==4.9
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.12
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.2
scour==0.38.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
Send2Trash==2.1.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
service-identity==24.2.0
setools==4.6.0
setproctitle==1.3.7
setuptools==78.1.1
shellingham==1.5.4
sherlock-project==0.15.0
shiboken6==6.10.2
simpleline==1.9.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.11.0
soupsieve==2.8
speg==0.3
SQLAlchemy==2.0.46
sqlparse==0.5.5
SSSDConfig==2.12.0
stack-data==0.6.3
starlette==0.52.1
stem==1.8.2
sympy==1.14.0
systemd-python==235
tables==3.10.2
tabulate==0.9.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
torbrowser-launcher==0.3.9
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
trio==0.32.0
truststore==0.10.4
Twisted==25.5.0
txaio==25.12.2
typer==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
typst==0.14.5
tzdata==2025.3
ufw==0.35
ujson==5.11.0
unicorn==2.1.2
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==14.2
Werkzeug==3.1.5
widgetsnbextension==4.0.15
wrapt==1.17.1
xarray==2025.12.0
xkbregistry==0.3
xlrd==2.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.3.3
zombie-imp==0.0.3
zope.interface==8.2
zstandard==0.25.0
